requirements.txt
click==6.7
Flask==0.12.2
Flask-JWT-Extended==3.1.0
Flask-Testing==0.6.2
Flask-WTF==0.14.2
itsdangerous==0.24
Jinja2==2.9.6
MarkupSafe==1.0
mongoengine==0.13.0
PyJWT==1.5.2
pymongo==3.4.0
six==1.10.0
Werkzeug==0.12.2
WTForms==2.1
pynacl==1.1.2